Refactor timed alignment into msa package
