+++ scinoephile/analysis/msa/__init__.py
@@ -1,0 +1,28 @@
+#  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
+#  and distributed under the terms of the BSD license. See the LICENSE file for details.
+"""Progressive multiple-sequence alignment of timestamped characters.
+
+Package hierarchy (modules may import from any above):
+* models
+* aligner / pauses
+"""
+
+from __future__ import annotations
+
+from .aligner import TimedMultiSequenceAligner
+from .models import (
+    TimedAlignmentColumn,
+    TimedAlignmentSequence,
+    TimedAlignmentSettings,
+    TimedAlignmentToken,
+    TimedMultiSequenceAlignment,
+)
+
+__all__ = [
+    "TimedAlignmentColumn",
+    "TimedAlignmentSequence",
+    "TimedAlignmentSettings",
+    "TimedAlignmentToken",
+    "TimedMultiSequenceAligner",
+    "TimedMultiSequenceAlignment",
+]
